move SGPR-specific test from test_methods to test_sgpr
